requirements.txt
pandas>=2.0.0
numpy>=1.24.0
biopython>=1.80
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.2.0
tqdm>=4.60.0
obonet>=1.0.0
networkx>=3.0